requirements.txt
pydantic-ai
pydantic
fastapi[standard]
pydantic-ai-slim[tavily]
markdown-pdf
markdown
slack-sdk
certifi
sqlmodel
alembic
python-dotenv